Add files via upload
This Machine Learning Model was built to predict the impact of differing advertising channels such as TV, radio and newspaper on sales in an organization. The model shows that every $1 spent on TV and radio will increase sale by 0.045 units and 0.188 units respectively. The model was tested by error metrics and the margin error was 10%. The model was later deployed by dump and load from joblib.